mahdi/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
mysql-connector-python
sqlalchemy
python-dotenv
langchain
langchain-groq
langchain-community
sentence-transformers
faiss-cpu
scipy
